do not start redis when on travis-ci

+++ spec/spec_helper.rb
@@ -9,37 +9,39 @@ $TESTING = true
 require 'minitest/autorun'
 require 'timecop'
 
-# make sure we can run redis
+if !ENV['TRAVIS']
 
-if !system("which redis-server")
-  puts '', "** can't find `redis-server` in your path"
-  puts "** try running `sudo rake install`"
-  abort ''
-end
+  # make sure we can run redis
 
-#
-# start our own redis when the tests start,
-# kill it when they end
-#
-
-at_exit do
-  next if $!
-
-  if defined?(MiniTest)
-    exit_code = MiniTest::Unit.new.run(ARGV)
-  else
-    exit_code = Test::Unit::AutoRunner.run
+  if !system("which redis-server")
+    puts '', "** can't find `redis-server` in your path"
+    puts "** try running `sudo rake install`"
+    abort ''
   end
 
-  pid = `ps -A -o pid,command | grep [r]edis-test`.split(" ")[0]
-  puts "Killing test redis server..."
-  `rm -f #{dir}/dump.rdb`
-  Process.kill("KILL", pid.to_i)
-  exit exit_code
+  #
+  # start our own redis when the tests start,
+  # kill it when they end
+  #
+
+  at_exit do
+    next if $!
+
+    if defined?(MiniTest)
+      exit_code = MiniTest::Unit.new.run(ARGV)
+    else
+      exit_code = Test::Unit::AutoRunner.run
+    end
+
+    pid = `ps -A -o pid,command | grep [r]edis-test`.split(" ")[0]
+    puts "Killing test redis server..."
+    `rm -f #{dir}/dump.rdb`
+    Process.kill("KILL", pid.to_i)
+    exit exit_code
+  end
+
+  puts "Starting redis for testing at localhost:9736..."
+  `redis-server #{dir}/redis-test.conf`
+  Widescreen.redis = 'localhost:9736'
 end
-
-puts "Starting redis for testing at localhost:9736..."
-`redis-server #{dir}/redis-test.conf`
-Widescreen.redis = 'localhost:9736'
-
